+++ CHANGELOG.md
@@ -10,9 +10,14 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 **Known Issue:** this version was compiled using typescript 3.7!
 The resulting `.d.ts` are **only** compatible with typescript versions `>= 3.6`. See [typescript 3.7 release notes](www.typescriptlang.org/docs/handbook/release-notes/typescript-3-7.html#class-field-mitigations) for more information.
 
+### Added
+
+- Added `currentViewWindow` to get the currently visible area of the graph
+
 ### Fixed
 
 - Fixed old node content staying when node type was changed
+- Fixed `dragend` event listed as `dragstart` in the documentation
 
 
 ## [0.4.1] - 2020-02-11